Accidentally made le64_t 32-bits wide... (I'm banging head on a wall right now)

+++ include/kanawha/endian.h
@@ -9,7 +9,7 @@ typedef uint16_t be16_t;
 typedef uint32_t le32_t;
 typedef uint32_t be32_t;
 
-typedef uint32_t le64_t;
-typedef uint32_t be64_t;
+typedef uint64_t le64_t;
+typedef uint64_t be64_t;
 
 #endif
